Move draw functions to utils - Note>this produces good beat
